Add utils.py
