+++ xinference/client/restful/async_restful_client.py
@@ -12,18 +12,17 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import json
-import typing
 from typing import TYPE_CHECKING, Any, AsyncIterator, Dict, List, Optional, Union
 
 import aiohttp
 
 from ..common import async_streaming_response_iterator
+from .utils import convert_float_to_int_or_str
 
 if TYPE_CHECKING:
     from ...types import (
         ChatCompletion,
         ChatCompletionChunk,
-        ChatCompletionMessage,
         Completion,
         CompletionChunk,
         Embedding,
@@ -31,17 +30,6 @@ if TYPE_CHECKING:
         PytorchGenerateConfig,
         VideoList,
     )
-
-
-def convert_float_to_int_or_str(model_size: float) -> Union[int, str]:
-    """convert float to int or string
-
-    if float can be presented as int, convert it to int, otherwise convert it to string
-    """
-    if int(model_size) == model_size:
-        return int(model_size)
-    else:
-        return str(model_size)
 
 
 def _filter_params(params: Dict[Any, Any]) -> Dict[Any, Any]:
@@ -74,25 +62,6 @@ async def _release_response(response: aiohttp.ClientResponse):
     """Release the aiohttp response."""
     response.release()
     await response.wait_for_close()
-
-
-@typing.no_type_check
-def handle_system_prompts(
-    chat_history: List["ChatCompletionMessage"], system_prompt: Optional[str]
-) -> List["ChatCompletionMessage"]:
-    history_system_prompts = [
-        ch["content"] for ch in chat_history if ch["role"] == "system"
-    ]
-    if system_prompt is not None:
-        history_system_prompts.append(system_prompt)
-
-    # remove all the system prompt in the chat_history
-    chat_history = list(filter(lambda x: x["role"] != "system", chat_history))
-    # insert all system prompts at the beginning
-    chat_history.insert(
-        0, {"role": "system", "content": ". ".join(history_system_prompts)}
-    )
-    return chat_history
 
 
 class AsyncRESTfulModelHandle:
